ml_breed_classifier/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
tensorflow==2.20.0
numpy>=2.1.0; python_version >= "3.13"
numpy>=1.26,<2; python_version < "3.13"
Pillow>=10.0.0
starlette==0.38.5
python-multipart==0.0.9
pandas>=1.5.0
scikit-learn>=1.3.0
kaggle>=1.6.14
